requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-dotenv
nltk
python-multipart
